+++ Example/StripePOS.framework/Headers/StripePOS.h
@@ -6,7 +6,6 @@
 //  Copyright © 2017 Stripe. All rights reserved.
 //
 
-#import "SCPAddress.h"
 #import "SCPBlocks.h"
 #import "SCPCancelable.h"
 #import "SCPCardBrand.h"