+++ StripePOS.framework/Headers/SCPDiscoveryDelegate.h
@@ -14,25 +14,28 @@
 NS_ASSUME_NONNULL_BEGIN
 
 /**
- Implement this protocol to handle events when connecting to a reader. You
- must pass a `SCPDiscoveryDelegate` when calling `connect`.
+ Implement this protocol to handle a continually updating list
+ of discovered readers. You must pass a `SCPDiscoveryDelegate`
+ when calling `discoverReaders`.
  */
 @protocol SCPDiscoveryDelegate <NSObject>
 
 /**
  This method will be called repeatedly until:
  - an error occurs
- - `discover` is canceled
+ - `discoverReaders` is canceled
  - `connect` is called with a discovered reader
 
- Your app should either display an updating list of discovered readers, or
- automatically select a previously saved reader. Once a selection has been made,
- call `SCPTerminal`'s `connect` method to begin connecting to the reader.
+ Your app should display an updating list of discovered readers if
+ your user is connecting to a reader for the first time.
+ Otherwise, you may automatically select a previously saved reader.
+ Once a selection has been made, call the `connect` method to begin
+ connecting to the reader.
 
  @param terminal        The originating SCPTerminal.
  @param readers         The discovered readers.
  */
-- (void)terminal:(SCPTerminal *)terminal didUpdateDiscoveryResults:(NSArray<SCPReader *>*)readers;
+- (void)terminal:(SCPTerminal *)terminal didUpdateDiscoveredReaders:(NSArray<SCPReader *>*)readers;
 
 @end